+++ ton_lib/src/cell/meta/cell_meta_builder.rs
@@ -52,7 +52,7 @@ impl<'a> CellMetaBuilder<'a> {
     }
 
     fn validate_ordinary(&self) -> Result<(), TonlibError> {
-        if self.data_bits_len as u32 > CellMeta::CELL_MAX_DATA_BITS_LEN {
+        if self.data_bits_len > CellMeta::CELL_MAX_DATA_BITS_LEN {
             return Err(TonlibError::BuilderMeta("Ordinary cell data bits length is too big".to_owned()));
         }
         Ok(())
